made delete confirmation dialog its own component
