+++ src/web/components/KeyPairs/KeyPairsTable.scss
@@ -4,7 +4,7 @@
   display: flex;
   flex-direction: column;
   align-items: stretch;
-  margin-bottom: 75px;
+  margin-bottom: 50px;
 
   .key-pairs-table {
     font-size: 0.875rem;
